refactor: split post and comment utils
